# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Dorra Alaseel - Complete API v1 Collection
  description: Complete API v1 endpoints for Dorra Alaseel Car Rental System - All
    /api/v1/ routes
  version: 1.0.0
servers:
  - url: "{{base_url}}"
...
paths:
...
  /api/v1/password/forgot/find/user:
    post:
      tags:
        - Authentication > User Auth
      summary: User Forgot Password - Find User
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              example: '"{\n  \"email\": \"user@example.com\"\n}"'
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/user/logout:
    post:
      tags:
        - User - Profile
      summary: Logout
      requestBody:
        content: {}
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{user_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/user/dashboard:
    get:
      tags:
        - User - Dashboard
      summary: Get Dashboard
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{user_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/user/notifications:
    get:
      tags:
        - User - Dashboard
      summary: Get Notifications
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{user_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/user/balance:
    get:
      tags:
        - User - Balance Management
      summary: Get Balance
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{user_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/user/branches:
    get:
      tags:
        - User - Branches & Delivery
      summary: List Branches
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{user_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/vendor/logout:
    post:
      tags:
        - Vendor - Profile
      summary: Logout
      requestBody:
        content: {}
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{vendor_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/vendor/dashboard:
    get:
      tags:
        - Vendor - Dashboard
      summary: Get Dashboard
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{vendor_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/vendor/notifications:
    get:
      tags:
        - Vendor - Dashboard
      summary: Get Notifications
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{vendor_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/vendor/due/pay:
    post:
      tags:
        - Vendor - Dashboard
      summary: Pay Due Amount
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              example: '"{\n  \"amount\": 1000\n}"'
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{vendor_token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/cars/vendor/{vendorId}:
    get:
      tags:
        - Cars (Public)
      summary: Get Cars by Vendor
      description: Get all cars from a specific vendor
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: sort
          in: query
          schema:
            type: string
          example: price_desc
        - name: vendorId
          in: path
          schema:
            type: integer
          required: true
          example: "5"
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
components:
...
  securitySchemes: {}